requirements.txt
dj-database-url==0.5.0
Django==2.0.7
django-classy-tags==0.8.0
django-cms==3.5.2
django-filer==1.3.2
django-filter==2.0.0
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==2.0.2
django-sekizai==0.10.0
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-column==1.8.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-installer==1.0.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
djangorestframework==3.8.2
easy-thumbnails==2.5
html5lib==0.9999999
Markdown==2.6.11
mysqlclient==1.3.13
Pillow==5.2.0
pytz==2018.5
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21
